Added visualizations
