+++ aermod.inp
@@ -5,7 +5,7 @@ CO STARTING
    AVERTIME  1  3  8  24  PERIOD
    POLLUTID  CACO3
    RUNORNOT  RUN
-   ERRORFIL  ERRORS.OUT
+   ERRORFIL  runs/ERRORS.OUT
 CO FINISHED
 
 SO STARTING
@@ -7790,17 +7790,17 @@ ME FINISHED
 OU STARTING
    RECTABLE  ALLAVE  FIRST
    MAXTABLE  ALLAVE  50
-   MAXIFILE  1  ALL  24.0  CACO3_100M_1HR_MAXIFILE.OUT
-   MAXIFILE  3  ALL  24.0  CACO3_100M_3HR_MAXIFILE.OUT
-   MAXIFILE  8  ALL  24.0  CACO3_100M_8HR_MAXIFILE.OUT
-   MAXIFILE  24  ALL  24.0  CACO3_100M_24HR_MAXIFILE.OUT
-   PLOTFILE  1  ALL  FIRST  CACO3_100M_1HR.PLT
-   PLOTFILE  3  ALL  FIRST  CACO3_100M_3HR.PLT
-   PLOTFILE  8  ALL  FIRST  CACO3_100M_8HR.PLT
-   PLOTFILE  24  ALL  FIRST  CACO3_100M_24HR.PLT
-   POSTFILE  1  ALL  PLOT  CACO3_100M_1HR.PST
-   POSTFILE  3  ALL  PLOT  CACO3_100M_3HR.PST
-   POSTFILE  8  ALL  PLOT  CACO3_100M_8HR.PST
-   POSTFILE  24  ALL  PLOT  CACO3_100M_24HR.PST
-   SUMMFILE  CACO3_100M_TEST.SUM
+   MAXIFILE  1  ALL  24.0  runs/CACO3_100M_1HR_MAXIFILE.OUT
+   MAXIFILE  3  ALL  24.0  runs/CACO3_100M_3HR_MAXIFILE.OUT
+   MAXIFILE  8  ALL  24.0  runs/CACO3_100M_8HR_MAXIFILE.OUT
+   MAXIFILE  24  ALL  24.0  runs/CACO3_100M_24HR_MAXIFILE.OUT
+   PLOTFILE  1  ALL  FIRST  runs/CACO3_100M_1HR.PLT
+   PLOTFILE  3  ALL  FIRST  runs/CACO3_100M_3HR.PLT
+   PLOTFILE  8  ALL  FIRST  runs/CACO3_100M_8HR.PLT
+   PLOTFILE  24  ALL  FIRST  runs/CACO3_100M_24HR.PLT
+   POSTFILE  1  ALL  PLOT  runs/CACO3_100M_1HR.PST
+   POSTFILE  3  ALL  PLOT  runs/CACO3_100M_3HR.PST
+   POSTFILE  8  ALL  PLOT  runs/CACO3_100M_8HR.PST
+   POSTFILE  24  ALL  PLOT  runs/CACO3_100M_24HR.PST
+   SUMMFILE  runs/CACO3_100M_TEST.SUM
 OU FINISHED
