+++ pyproject.toml
@@ -5,6 +5,8 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
+    "imageio>=2.37.0",
     "matplotlib>=3.10.3",
     "numpy>=2.3.1",
+    "pandas>=2.3.1",
 ]
